requirements.txt
groq
cohere
rich
python-dotenv
google-generativeai
crewai[tools]
crewai